pyproject.toml
[tool.poetry]
name = "scribes"
version = "0.0.0"
description = ""
authors = ["Armand Sauzay <35524799+armand-sauzay@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7"
click = ">=8.1.7"
rich = ">=13.6.0"
pyyaml = ">=6.0.1"
types-pyyaml = ">=6.0.12.12"


[tool.poetry.group.test.dependencies]
pytest = ">=7.4.2"
pytest-mock = ">=3.11.1"

[tool.poetry.scripts]
scribes = "scribes.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
